+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/RoomAvatarResolver.kt
@@ -20,14 +20,13 @@ import com.zhuinden.monarchy.Monarchy
 import im.vector.matrix.android.api.auth.data.Credentials
 import im.vector.matrix.android.api.session.events.model.EventType
 import im.vector.matrix.android.api.session.events.model.toModel
-import im.vector.matrix.android.api.session.room.model.Membership
 import im.vector.matrix.android.api.session.room.model.RoomAvatarContent
+import im.vector.matrix.android.api.session.room.model.RoomMember
 import im.vector.matrix.android.internal.database.mapper.asDomain
 import im.vector.matrix.android.internal.database.model.EventEntity
-import im.vector.matrix.android.internal.database.model.RoomEntity
+import im.vector.matrix.android.internal.database.model.EventEntityFields
 import im.vector.matrix.android.internal.database.query.prev
 import im.vector.matrix.android.internal.database.query.where
-import im.vector.matrix.android.internal.session.SessionScope
 import im.vector.matrix.android.internal.session.room.membership.RoomMembers
 import javax.inject.Inject
 
@@ -42,32 +41,25 @@ internal class RoomAvatarResolver @Inject constructor(private val monarchy: Mona
     fun resolve(roomId: String): String? {
         var res: String? = null
         monarchy.doWithRealm { realm ->
-            val roomEntity = RoomEntity.where(realm, roomId).findFirst()
             val roomName = EventEntity.where(realm, roomId, EventType.STATE_ROOM_AVATAR).prev()?.asDomain()
             res = roomName?.content.toModel<RoomAvatarContent>()?.avatarUrl
             if (!res.isNullOrEmpty()) {
                 return@doWithRealm
             }
             val roomMembers = RoomMembers(realm, roomId)
-            val members = roomMembers.getLoaded()
-            if (roomEntity?.membership == Membership.INVITE) {
-                if (members.size == 1) {
-                    res = members.entries.first().value.avatarUrl
-                } else if (members.size > 1) {
-                    val firstOtherMember = members.filterKeys { it != credentials.userId }.values.firstOrNull()
-                    res = firstOtherMember?.avatarUrl
-                }
-            } else {
-                // detect if it is a room with no more than 2 members (i.e. an alone or a 1:1 chat)
-                if (members.size == 1) {
-                    res = members.entries.first().value.avatarUrl
-                } else if (members.size == 2) {
-                    val firstOtherMember = members.filterKeys { it != credentials.userId }.values.firstOrNull()
-                    res = firstOtherMember?.avatarUrl
-                }
+            val members = roomMembers.queryRoomMembersEvent().findAll()
+            // detect if it is a room with no more than 2 members (i.e. an alone or a 1:1 chat)
+            if (members.size == 1) {
+                res = members.firstOrNull()?.toRoomMember()?.avatarUrl
+            } else if (members.size == 2) {
+                val firstOtherMember = members.where().notEqualTo(EventEntityFields.STATE_KEY, credentials.userId).findFirst()
+                res = firstOtherMember?.toRoomMember()?.avatarUrl
             }
-
         }
         return res
+    }
+
+    private fun EventEntity?.toRoomMember(): RoomMember? {
+        return this?.asDomain()?.content?.toModel<RoomMember>()
     }
 }