+++ vector/src/main/java/im/vector/riotx/features/home/room/list/RoomListFragment.kt
@@ -38,6 +38,7 @@ import im.vector.riotx.core.platform.OnBackPressed
 import im.vector.riotx.core.platform.StateView
 import im.vector.riotx.core.platform.VectorBaseFragment
 import im.vector.riotx.features.home.room.list.widget.FabMenuView
+import im.vector.riotx.features.notifications.NotificationDrawerManager
 import kotlinx.android.parcel.Parcelize
 import kotlinx.android.synthetic.main.fragment_room_list.*
 import javax.inject.Inject
@@ -69,6 +70,7 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
     @Inject lateinit var roomController: RoomSummaryController
     @Inject lateinit var roomListViewModelFactory: RoomListViewModel.Factory
     @Inject lateinit var errorFormatter: ErrorFormatter
+    @Inject lateinit var notificationDrawerManager: NotificationDrawerManager
     private val roomListViewModel: RoomListViewModel by fragmentViewModel()
 
     override fun getLayoutResId() = R.layout.fragment_room_list
@@ -112,7 +114,7 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
         }
 
         // Hide FAB when list is scrolling
-        epoxyRecyclerView.addOnScrollListener(
+        roomListEpoxyRecyclerView.addOnScrollListener(
                 object : RecyclerView.OnScrollListener() {
                     override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
                         createChatFabMenu.removeCallbacks(showFabRunnable)
@@ -136,11 +138,14 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
     }
 
     fun filterRoomsWith(filter: String) {
+        // Scroll the list to top
+        roomListEpoxyRecyclerView.scrollToPosition(0)
+
         roomListViewModel.accept(RoomListActions.FilterWith(filter))
     }
 
-    override fun openRoomDirectory() {
-        navigator.openRoomDirectory(requireActivity())
+    override fun openRoomDirectory(initialFilter: String) {
+        navigator.openRoomDirectory(requireActivity(), initialFilter)
     }
 
     override fun createDirectChat() {
@@ -150,12 +155,12 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
     private fun setupRecyclerView() {
         val layoutManager = LinearLayoutManager(context)
         val stateRestorer = LayoutManagerStateRestorer(layoutManager).register()
-        epoxyRecyclerView.layoutManager = layoutManager
-        epoxyRecyclerView.itemAnimator = RoomListAnimator()
+        roomListEpoxyRecyclerView.layoutManager = layoutManager
+        roomListEpoxyRecyclerView.itemAnimator = RoomListAnimator()
         roomController.listener = this
         roomController.addModelBuildListener { it.dispatchTo(stateRestorer) }
-        stateView.contentView = epoxyRecyclerView
-        epoxyRecyclerView.setController(roomController)
+        stateView.contentView = roomListEpoxyRecyclerView
+        roomListEpoxyRecyclerView.setController(roomController)
     }
 
     private val showFabRunnable = Runnable {
@@ -255,10 +260,12 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
     }
 
     override fun onAcceptRoomInvitation(room: RoomSummary) {
+        notificationDrawerManager.clearMemberShipNotificationForRoom(room.roomId)
         roomListViewModel.accept(RoomListActions.AcceptInvitation(room))
     }
 
     override fun onRejectRoomInvitation(room: RoomSummary) {
+        notificationDrawerManager.clearMemberShipNotificationForRoom(room.roomId)
         roomListViewModel.accept(RoomListActions.RejectInvitation(room))
     }
 
@@ -266,9 +273,8 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
         roomListViewModel.accept(RoomListActions.ToggleCategory(roomCategory))
     }
 
-    // TODO Pass title
-    override fun createRoom() {
-        navigator.openCreateRoom(requireActivity())
+    override fun createRoom(initialName: String) {
+        navigator.openCreateRoom(requireActivity(), initialName)
     }
 
 }