+++ CHANGES.md
@@ -2,19 +2,25 @@ Changes in RiotX 0.2.1 (2019-XX-XX)
 ===================================================
 
 Features:
- - Message Editing: View edit history
+ - Message Editing: View edit history (#121)
  - Rooms filtering (#304)
 
 Improvements:
  - Handle click on redacted events: view source and create permalink
+ - Improve long tap menu: reply on top, more compact (#368)
+ - Quick reply in timeline with swipe gesture (#167)
+ - Improve edit of replies
+ - Improve performance on Room Members and Users management (#381)
 
 Other changes:
- -
+ - migrate from rxbinding 2 to rxbinding 3
 
 Bugfix:
  - Fix regression on permalink click
  - Fix crash reported by the PlayStore (#341)
  - Fix Chat composer separator color in dark/black theme
+ - Fix bad layout for room directory filter (#349)
+ - Fix Copying link from a message shouldn't open context menu (#364)
 
 Translations:
  -